+++ Changes.md
@@ -132,7 +132,7 @@ Breaking Changes
   - MeshToLevelSet
   - LevelSetOffset
   - PointsGridToPoints
-
+- LevelSetToMesh : Changed default value for `adjustBounds` plug.
 
 Build
 -----